+++ docs/snippets/tutorial/season_detrend_v3.py
@@ -1,12 +1,12 @@
 from season_seismogram import SeasonSeismogram, Season
 from functions_v4 import detrend
-from datetime import datetime
+from pandas import Timestamp
 import numpy as np
 
 
 def main() -> None:
     # Create a sample SeasonSeismogram instance with random data
-    begin_time = datetime(2023, 1, 1, 0, 0, 0)
+    begin_time = Timestamp(2023, 1, 1, 0, 0, 0)
     data = np.random.randn(1000)
     season_seismogram = SeasonSeismogram(
         begin_time=begin_time, data=data, season=Season.WINTER