+++ tests/tools/test_plotutils.py
@@ -1,5 +1,7 @@
 from pysmo import Seismogram
 from pytest_cases import parametrize_with_cases
+from pandas import Timestamp
+from datetime import timezone
 import pytest
 import matplotlib  # type: ignore
 
@@ -23,19 +25,28 @@ class TestPlotseisFunctions:
 
         times = time_array(seismogram)
         assert len(times) == len(seismogram)
-        assert num2date(times[0]) == seismogram.begin_time
-        assert num2date(times[-1]) == seismogram.end_time
+        assert (
+            Timestamp(num2date(times[0])).timestamp()
+            == seismogram.begin_time.timestamp()
+        )
+        assert (
+            Timestamp(num2date(times[-1])).timestamp()
+            == seismogram.end_time.timestamp()
+        )
 
     def test_unix_time_array(self, seismogram: Seismogram) -> None:
         """Get times from Seismogram object and verify them."""
         from pysmo.tools.plotutils import unix_time_array
-        from datetime import datetime, timezone
 
         unix_times = unix_time_array(seismogram)
         assert len(unix_times) == len(seismogram)
         assert (
-            datetime.fromtimestamp(unix_times[0], timezone.utc) == seismogram.begin_time
+            pytest.approx(
+                Timestamp.fromtimestamp(unix_times[0], timezone.utc).timestamp()
+            )
+            == seismogram.begin_time.timestamp()
         )
         assert (
-            datetime.fromtimestamp(unix_times[-1], timezone.utc) == seismogram.end_time
+            Timestamp.fromtimestamp(unix_times[-1], timezone.utc).timestamp()
+            == seismogram.end_time.timestamp()
         )